requirements.txt
# Core UI
streamlit==1.51.0

# LLM & Orchestration
langchain-core==1.2.5
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-groq==1.1.1
langchain-tavily==0.2.17
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.2

# Vector Search (RAG)
faiss-cpu==1.12.0

# Embeddings & Utilities
langchain-huggingface==1.0.1

# Reranking
# sentence-transformers==5.1.2

# Environment Variables
python-dotenv==1.2.1

# Data Validation
pydantic==2.12.3

# PDF Processing (if used)
pypdf==6.5.0
